+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This document logs the changes per release of TWCManager.
   * Added ability to override the policy check timer to set it to any value desired
   * Added ability to override the policy entirely from the configuration file (with caveat that this is not recommended)
   * Added ability to extend the policy at two points (before and after) to add extra rules in whilst retaining the rest of the policy
+  * MQTT control command chargeNow will now accept a payload specifying the number of amps and duration to charge
   * Bugfixes
      * Fixed an issue where the policy based charging rules could not access settings variables due to an error in string offset (thanks @AndySchroder).
      * Fixed a bug with debug output if the charger is configured to draw more amps than the wiring can handle (thanks @AndySchroder).